Decode int32 proto fields with Int32 instead of Int64
datasize, raw_size, granularity, and date_granularity are int32 in the
PBF schema. Decoding them as Int64 then casting worked but obscured the
proto types. Switch to Int32 so the call site matches the schema.

+++ block.go
@@ -103,12 +103,12 @@ func (br *BlockReader) Next() bool {
 				br.blockType = string(b)
 			}
 		case 3:
-			v, err := hMsg.Int64()
+			v, err := hMsg.Int32()
 			if err != nil {
 				br.err = fmt.Errorf("osmbr: BlobHeader.datasize: %w", err)
 				return false
 			}
-			dataSize = v
+			dataSize = int64(v)
 		default:
 			hMsg.Skip()
 		}
